requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
Pillow==10.4.0
PyYAML==6.0.2
aiofiles==24.1.0
python-multipart==0.0.9
